+++ src/main/java/com/q6cyber/assessment/service/ForecastChecker.java
@@ -1,6 +1,7 @@
-package com.q6cyber.assessment;
+package com.q6cyber.assessment.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.q6cyber.assessment.Config;
 import com.q6cyber.assessment.api.geocode.GeocodeResponse;
 import com.q6cyber.assessment.api.weather.ForecastGridResponse;
 import com.q6cyber.assessment.api.weather.ForecastResponse;
